pyproject.toml
[tool.poetry]
name = "sendgrid-async"
version = "2.0.4"
description = "SendGrid using an httpx client"
license = "MIT"
authors = ["Elyes Mahjoubi <elyesmahjoubi@gmail.com>"]
readme = "README.md"
repository = "https://github.com/EM51641/async-sendgrid-"
keywords = ["sendgrid", "client", "async"]
classifiers = [
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

packages = [
    { include = "async_sendgrid" },
]

include = [
    "LICENSE",
    "README.md"
]

[tool.poetry.dependencies]
python = "^3.10"
sendgrid = "^6.7.0"
httpx = "^0.24.1"
opentelemetry-api = "^1.34.0"
opentelemetry-sdk = "^1.34.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.34.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.15.1"
pytest-cov = "^4.0.0"
mypy = "^1.0.0"
flake8 = "^6.0.0"
black = "^23.9.0"

[tool.black]
line-length = 79

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-api>=1.20.0",
]

[tool.pytest-asyncio]
asyncio_mode = "auto"